+++ src/test/java/org/remipassmoilesel/bookme/MerchantServiceControllerTest.java
@@ -104,7 +104,7 @@ public class MerchantServiceControllerTest {
         for (int i = 0; i < servicesNumber; i++) {
             MerchantServiceType srvT = (MerchantServiceType) Utils.randValueFrom(serviceTypes);
             Customer customer = (Customer) Utils.randValueFrom(customers);
-            int price = srvT.getPrice() + i;
+            double price = srvT.getPrice() + i;
             String comment = Utils.generateLoremIpsum(100);
             MerchantService srv = new MerchantService(srvT, customer, price, comment,
                     beginTestPeriod.plusDays(i).toDate(),
@@ -116,7 +116,7 @@ public class MerchantServiceControllerTest {
         for (int i = 0; i < servicesNumber; i++) {
             MerchantServiceType srvT = (MerchantServiceType) Utils.randValueFrom(serviceTypes);
             Customer customer = (Customer) Utils.randValueFrom(customers);
-            int price = srvT.getPrice() + i;
+            double price = srvT.getPrice() + i;
             String comment = Utils.generateLoremIpsum(100);
             MerchantService srv = new MerchantService(srvT, customer, price, comment,
                     beginTestPeriod.plusDays(i).toDate(),
